packages/bridge/requirements.txt
fastapi==0.118.2
uvicorn[standard]==0.24.0
sqlalchemy==2.0.32
psycopg[binary]==3.3.2
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-dotenv==1.0.0
pydantic==2.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
alembic==1.13.1
httpx==0.25.2